requirements.txt
google-cloud-aiplatform==1.82.0
google-cloud-translate==3.20.1
pandas==2.2.3
tenacity==9.0.0
